create dashboard Fixes #12
