+++ e-commerce-react-vite-app/src/components/data/inventory_data_types.tsx
@@ -1,0 +1,7 @@
+export type InventoryData = {
+    // Each attribute name must match the header name from the csv file
+    nombre: string;
+    precio_con_tax: number;
+    precio_enviado: number;
+    ganancia_para_envio: number;
+}
