+++ e-commerce-react-vite-app/src/components/helper_functions/FetchCSVData.tsx
