+++ e-commerce-react-vite-app/src/components/inventory_dashboard/InventoryDashBoard.tsx
@@ -1,0 +1,8 @@
+import { InventoryDashBoardProps } from "./types"
+
+export const InventoryDashBoard = ({total_spent}: InventoryDashBoardProps) => {
+    return (
+        <div>InventoryDashBoard Total Spent: {total_spent}</div>
+    )
+}
+
