+++ e-commerce-react-vite-app/src/components/inventory_dashboard/types.tsx
@@ -1,0 +1,3 @@
+export type InventoryDashBoardProps = {
+    total_spent: number;
+}
