+++ e-commerce-react-vite-app/src/components/inventory_table/InventoryTableRenderer.tsx
@@ -1,20 +1,8 @@
 import Table from "../table/Table";
-import { fetchCSVData } from "../data/FetchCSVData";
-import { InventoryData, InventoryColumns, InventoryTableRendererProps } from "./types";
-import { useEffect, useState } from "react";
+import { InventoryColumns, InventoryTableRendererProps } from "./types";
 import InventoryTableTitleBar from "./InventoryTableTittleBar";
-import { inventoryColumnSums } from "../data/DataProcessor";
 
-export default function InventoryTableRenderer({ date, data_source_link }: InventoryTableRendererProps) {
-    const [inventoryData, setInventoryData] = useState<InventoryData[]>([]);
-
-    useEffect(() => {
-        fetchCSVData<InventoryData>(data_source_link)
-            .then(data => setInventoryData(data));
-    }, [data_source_link]);
-
-    const { totalSpent, totalProfitForSending } = inventoryColumnSums(inventoryData);
-
+export default function InventoryTableRenderer({ date, totalSpent, totalProfitForSending, inventoryData }: InventoryTableRendererProps) {
     return (
         <div>
             <div>
