+++ e-commerce-react-vite-app/src/components/inventory_table/types.tsx
@@ -1,12 +1,5 @@
 import { Column } from "../table/interfaces";
-
-export type InventoryData = {
-    // Each attribute name must match the header name from the csv file
-    nombre: string;
-    precio_con_tax: number;
-    precio_enviado: number;
-    ganancia_para_envio: number;
-}
+import { InventoryData } from "../data/inventory_data_types";
 
 export const InventoryColumns: Column<InventoryData>[] = [
     { header: "Nombre", accessor: "nombre" },
@@ -15,7 +8,6 @@ export const InventoryColumns: Column<InventoryData>[] = [
     { header: "Ganancia para Envio", accessor: "ganancia_para_envio" },
 ];
 
-
 export type InventoryTableTitleBarProps = {
     date: string;
     total_spent: number;
@@ -23,6 +15,8 @@ export type InventoryTableTitleBarProps = {
 }
 
 export type InventoryTableRendererProps = {
-    date: string
-    data_source_link: string
+    date: string;
+    totalSpent: number;
+    totalProfitForSending: number;
+    inventoryData: InventoryData[];
 }
